refactor: rename errorpage
